jackpot-ingest/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
psycopg2-binary==2.9.9
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
python-dateutil==2.9.0
